+++ Runtime/Core/AssemblyInfo.cs
@@ -1,3 +1,4 @@
 using System.Runtime.CompilerServices;
 
 [assembly: InternalsVisibleTo("Deucarian.Notifications.Tests")]
+[assembly: InternalsVisibleTo("Deucarian.Notifications.Editor")]